Fix the bug introduced into this copy/pasted version of ultoac (unsigned-long to ASCII with commas) in rev 1.68 (addressing warnings). A size_t (i) can't go negative. Duh.

+++ src/sbbs3/chksmb.c
@@ -55,7 +55,7 @@
 char *ultoac(ulong l, char *string)
 {
 	char str[256];
-	size_t i,j,k;
+	int i,j,k;
 
 	SAFEPRINTF(str,"%lu",l);
 	i=strlen(str)-1;
